First alpha revision
:file:`Proposition-ChatGPT.css`: CSS code proposed by ChatGPT in course
of approaching the problem.

:file:`html4css1.css`: The main CSS used by docutils.  It is noteworthy
that neither size nor spacing of headlines is declared by this file.  So
I am going to reproduce the default setting of my browser.

:file:`Native.{rst, html}`: Used to inspect the rendering of
:file:`Native.rst` in :file:`Native.html`, Created by just ``$ rst2html
Native.rst Native.html``.

:file:`docutils.css`: The CSS written by me.  It currently:

-   Uses ``font-family: serif`` in the body of an article, in its
    filename display (without extension) just above the body, and in the
    preview.

-   Styles headline based on figures provided by ChatGPT, both in
    preview and live mode, by applying ``padding``, symmetrical
    top/bottom.

-   Eliminates an almost empty ``div.cm-line`` inserted by Obsidian just
    above tables.

The heading spacing will be adjusted based on further research carried
out yesterday.

An adjustment of font size in ``<tt>`` elements is required, here the
choice of docutils will be investigated.

A separation of text paragraphs by ``1em`` *in the live mode* is
desirable.

Some tidying up might be possible.