Updated padding rules; corrected a fix regarding tables
I applied new padding rules.  These *only* set ``padding-top:``.  The
choices are still not optimal for Obsidian rendering.  I will need to
reinvestigate this.

In::

    .div.HyperMD-header.cm-line + .cm-line { padding-top: 1.11em !important; }

the ``!important`` is required to override an Obsidian choice.  This in
turn makes ``!important`` required also in the following rules, which
have the same speicificity.

The selector in the rule to trim away the empty Obsidian-caused empty
line preceding tables has been improved.  Before, it matched *any* empty
line containing only a ``<br>``, which removed *all* of them; the new
selector now only selects those empty lines which truly precede a table.
There is a little flaw: The cursor can be positioned in the trimmed
line.  This seems unfixable.