Reiterated the spacing around headers; improved separator lines CSS
The ``padding-top`` figures of the headers have been recalucated using
measurements directly from within Obsidian.  Things appeat to be a lot
more pleasant now.

Improved the rule for empty lines *between paragraphs*: The selector now
expects ``:has(+ .cm-line)``.  This prevents the selector to match the
*last line* of input.  Without this selector component, the cursor
reduced to the 40% size when it was in the last line of input, before
the user begins to type content.  Now, the last line is kept at the 100%
size, improving look-and-feel.

*Both* rules regarding empty lines have been improved, by removing the
``padding`` and ``pargin`` declaration, these are not needed.