Reducing font size for element showing code
Inline code sequences are now scaled to 75% font size.  This is what I
found in Firefox as the default style.

Code sequences which are *not* inline use 85% font size.  The figure is
my own reasoning, attempting a pleasant impression.

I reorganised the overall structure of the CSS code a little (moving
things around and adding comments).