Further improving the rendering of empty lines
Now, *all* empty lines are rendered *per default* with ``font-size:
40%``, however, for those which carry the ``cm-active`` class,
``font-size`` is set back to ``100%``.  This ensures that all empty
lines which carry the cursor are rendered in full height.

Before, such empty lines were rendered in full size only when they were
the last line of the document, independent on whether the cursor was
inside of them or not.  This resulted in the desired effect during
writing the document, when the cursor was at its end, but it did *not*
change the height of inner empty lines, leading to a cursor reduced in
size prior to first input on that line.

The new code has the drawback, that the increased temporary height
shifts the text following it downwards, but this I can cope with; to my
impression it is more important that the cursor is in full height *from
the beginning*.