Headlines; overall *serif*; whitespace around lists; inline title
The default headline variables (size, weight, variant) are now set to
the values extracted from the default theme.  This overrides choices by
the Minimal theme.

The code to choose a serif font in the inline title and the editor pane
has been changed:

1.  For the *preview mode*, it is sufficient to declare ``--font-text:
    serif;``, both when testing with the default theme as well as with
    the Minimal theme.

2.  For Live Mode, I am now selecting ``div.cm-editor``, and style it
    with *both* ``--font-text: serif;`` *as well as* ``font-family:
    serif;``.

    I observed, that the ``--font-text`` declaration resulted *only in
    the default theme* in a serif font, while ``font-family: serif;``
    only had this effect in the Minimal theme; I cannot explain this
    behaviour.

Two variants for injecting whitespace related to lists have been worked
out:

1.  Adding margin around *each* list item (variant *A*);
2.  Injecting whitespace *only around entire* lists (variant *B*).

While (A) is strictly speaking typographically more correct, it is
unusual in the context of ReST/MD formatting.  I therefore enabled
variant (B).  (A) remains in the code commented-out.

Overriding a nonzero ``margin-bottom`` for the ``div.inline-title``.

Now, live mode and preview mode are close to identical in rendering.