Refined the live header CSS codes
The implementation of whitespace around headlines as been changed:

-   Headline produce, as before, padding above only.  This leads to the
    proper whitespace between elements of a headline stack.  The
    whitespace *below* such a stack is provided by the first line
    following it.

-   However, this led to problems when the document started by a
    headline.  Its padding added up graphically with the margin of the
    inline title.

-   I am now avoiding introduction of ``padding-top`` whitespace *by the
    first headline* when this is the first document line (``.cm-line``).

-   The margin of the inline title has been set to the same values as
    used for body ``h1`` elements.

-   For the first headline, when it has no ``.cm-line`` above it,
    ``padding-top`` is set zero.  In all other cases, the
    ``padding-top`` is applied as before.

In this way, the *first* space just above the document is provided by
the inline title, while all headline whitespace *within* the document is
handled by the CSS concerning the editor.

As an equivalent alternative to this approach, a strategy using a
``:first-child`` selector might be conceivable.

Before, the first headline in the document distanced itself to the top
by a ``margin: xxem 0 !important;`` declaration.  This space didn't
collapse with the margin of the inline title, leading to a typographic
flaw.