Refined :file:`Checklists.css`
I *commented out* the declaration decorating the text of the *cancelled*
('-') checkmark by a strike-through line.  However, the ``color:``
declaration is now in effect again.

Notice that checked checkmarks use *by default* a slightly greyed out
text colour as well.

The comment at the top of the file has been updated and corrected (it
referred to the "partial" checkmark icon where the "cancelled" one is
meant.