Synchronising ``font-size`` in blank table-predecessor lines
Tables are inevitably preceded by a blank line.  The ``height`` of these
blank lines is already reduced to ``0``.  Now, the ``font-size`` is
reduced from 40% to 30% as it was implemented for ordinary blank lines
already.  It only affects the cursor size.  This change is rather
cosmetic and has very little impact.