Setting the line height
Specifying ``--line-height-normal`` as ``1.3``.  Before, both with the
default theme as well as with Minimal, it was conistently set to
``1.5``.  Especially on the iPad however this is probabaly too
pronounced.

It might be, that serif fonts are more susceptible to increased line
height.  Maybe ``1.3`` would be on the contrary *too small* with a
sans-serif font.