Some editing of :file:`0300 Installation.md`
Incorporating the former section "The CSS Snippets" into the first
section, which did not have a title before and now is called "Contents
of the Repository".

*Editied in* the Finder menu item description for the procedure on
macOS.

Added small sections "Windows" and "Linux".  On Windows, a usage scheme
of the Explorer is described.  On Linux, almost nothing can be said.