Adding a comment about a ``font-size: 38%`` declaration
Added a comment on the rule selecting the blank line preceding tables.
I played with the idea of removing the ``font-size: 38%`` declaration,
but then it would use the ``font-size: 100%`` from the ``.cm-active``
rule.  In that case, the cursor line reaches downwards to the first row
of the table.  Some *other* value, like ``50%``, would do as well, but
I'd like to keep the values consistent.