Introducing --font-size-blank-line and --margin-blank-line
I noticed that the *vertical margin* for paragraphs and lists was set to
``0.4em`` statically.  This value led to whitespace differing a little
in reading mode from the value used (for blanks) in editing mode.

To make both results (in editing and reading mode) looks as identical as
possible, the reading mode margin corresponding to blank lines in
editing mode needs to be calculated from 1) the *font size* of such
blank lines and 3) the line height of the text where it occurs in.  This
calculation has been carried out and the resulting rendering is as
independent on the mode as it seems possible.

To avoid pitfalls from needing to change this settings *in several
places* at the same time, there are now central *variables*
``--font-size-blank-line`` and ``margin-blank-line``, including a
comment describing the calculation needed to determine
``--margin-blank-line``.  These vars are defined in the same place as
``--line-height-normal``.