Propagated changes to the :file:`Typo-Doc`
I noticed a limitation of *Obsisian-Typo*: When formatting an .md
document with *intentional blank lines* between items of a list, these
blanks will be *ignored* when moving from editing mode to reading mode.

There's not much to help with this.  The blank lines *need to be shown*
in editing mode, and at the same time *need to be ignored* in preview
rendering.