Exempting codeblock lines from emty-paragraphs styling
The selectors for squeezing empty lines to
``var(--font-size-blank-line)`` and for re-setting the font-size of
empty lines *with the cursor* to 100% now *only* apply to empty
``.cm-lines`` which are *not* ``.HyperMD-codeblock`` lines.

For *empty code block lines* which *carried the cursor*, the font size
was set to 100% by the rule intended to re-set font size.  This was
larger than the 75% set by the code block rule.  As a result, the height
of the line increased slightly when visiting it with the cursor,
shifting contents slightly.

For empty code block lines *not* carrying the cursor, the font size was
set to 75%.  Exempting code block lines from the selector of the rule to
squeeze down empty lines is therefore not strictly required.