Using the ``--code-size`` variable to style codes
There exist rule by Obsidian implementing ``--code-size``, so with the
declaration ``--code-size: 75%`` which now is provided, my own rules to
select code blocks and inline code are no longer needed.

However, for inline codes in headers, we need to override a declaration
from Obsidian ``font-size: inherit !important`` from Obsidian, which is
because of the ``!important`` overriding the abovementioned rules.
Here, we install ``font-size: var(--code-size) !important``.