Code blocks directly beneath a headline
Fixed a glitch when code blocks were placed directly beneath a headline.
In this case, the first code block line was given ``1.00em`` of
paddding, however, this is measured in --code-size lines.  Thus, code
block lines need to be provided with a padding whose size neutralises
the smaller font height; a new variable ``--code-1em`` was added for
this purpose.

With the old code, the problem is reproducible by this checklist:

1.  Write a headline.
2.  Write a code block beneath it.
3.  Add a blank line in front of the code block, between the code block
    and the headline.

After (2.), the first code block line will receive additional padding.
By (3.), this padding is replace by the whitespace of the blank line.

From (2.) to (3.) the code content of the code block was shifted a few
px.  With the new codes, this no longer occurs.

Writing code blocks directly beneath a headline is not recommended,
because the padding between its content and the headline will be
"greyspace", not "whitespace", depending on the background colour of the
code block.  To circumvent this, equivalent MD content is
recommended, with a blank between the code block and the headline.  Now,
following this recommendation no longer leads to the abovementioned
glitch, making applying the "fix" more smooth for the user.