Improved the code dealing with *empty lines*
With the previous code, the rule requesting --font-size-blank-line for
empty lines preceding a table didn't override the ``font-size``, only
the ``height: 0 !important`` was effective.  This was due to a higher
specificity of the rule selecting the ``.cm-active`` blank line.

I tuned the rule for the ``.cm-active`` element to no longer select it
when it *is* followed by a table.  By this change, there also is no need
any longer to request ``--font-size-blank-line`` for such empty lines;
this will be covered by the general rule applied to blank lines.