Documentation ready for proof-reading; bugfix in :file:`Typo-Core.css`
The Sect. "Notes on the Implementation" has seen substantial amount of
work.  It currently consist of a near-to-stable section, followed after
an horizontal line by remnants of the approach to the current production
contents.

In :file:`Typo-Core.css`, a typo in the ``--h4-size`` variable has been
spotted: It ran before ``1.118em``, where the current, correct measure
is ``1.188em``.  This last value stems from revisiting headlines in the
default theme by the Obsidian inspector, and it matches perfectly the
``padding-top`` for level-4 headers already in action.

The stylesheets of the :file:`Typo-Doc` vault has been updated to the
current versions.

I will now prune in the "Implementation" chapter the remnant texts
(below the horizontal line) and will apply proof-reading.

Intended is also a section on how to *test* Obsidian-Typo, by describing
ways of interacting with Obsidian while the Typo styles are in effect.