Further tweaks
Numerous small fixes.

Added the *Notes on the Implementation* document to the ToC.

In the *Installation* chapter, the *you* in the introductory text on the
*Procedure* has been avoided.  It is maintained in the checklists.

The description of the exeptions from rending blank lines reduced in
size are now enumerated and have been transformed into dedicated
paragraphs.

A few points on how to define whitespace figures have been incorporated
into the list of the followed procedure.